pyproject.toml
[tool.poetry]
name = "happy-birthday"
version = "0.1.0"
description = ""
authors = ["grbcas <grbcas@gmail.com>"]
readme = "README.md"
packages = [{include = "happy_birthday"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.3"
python-dotenv = "^1.0.1"
djangorestframework = "^3.15.1"
celery = "^5.4.0"
requests = "^2.31.0"
django-celery-beat = "^2.6.0"
psycopg2-binary = "^2.9.9"
django-extensions = "^3.2.3"
redis = "^5.0.4"
flower = "^2.0.1"
gunicorn = "^22.0.0"
ipython = "^8.25.0"
django-debug-toolbar = "^4.4.2"
djangorestframework-simplejwt = "^5.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
